pyproject.toml
[project]
name = "books-generation-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "google-genai>=1.72.0",
    "gradio>=6.12.0",
    "openpyxl>=3.1.5",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.2",
    "sqlalchemy>=2.0.49",
]
